requirements.txt
-i https://pypi.org/simple
fastapi==0.60.1
google-api-core[grpc]==1.22.2
google-auth==1.21.2
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
joblib==0.16.0
matplotlib
nltk==3.4.4
opencv-python
pandas==1.1.0
pep8
plotly==4.9.0
pydantic
pytesseract
python-dotenv==0.15.0
requests==2.24.0
scikit-image
scikit-learn==0.23.2
sqlalchemy==1.3.20
uvicorn
wordcloud==1.8.1
